correct validation api

+++ api/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -9,10 +10,15 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid User Id' });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
-      return response.status(404).json({ message: 'user not found' });
+      return response.status(404).json({ error: 'Contact not found' });
     }
     response.status(200).json(contact);
   }
@@ -21,45 +27,68 @@ class ContactController {
     const {
       name, email, phone, category_id,
     } = request.body;
-    if (!name) {
-      return response.status(404).json({ message: 'name is required' });
-    }
-    const contactExists = await ContactsRepository.findByEmail(email);
 
-    if (contactExists) {
-      return response.status(404).json({ message: 'This e-mail address is already in use' });
+    if (!name) {
+      return response.status(404).json({ error: 'name is required' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
+    }
+
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if (contactExists) {
+        return response.status(404).json({ error: 'This e-mail address is already in use' });
+      }
     }
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
     response.status(201).json(contact);
   }
 
   async update(request, response) {
     const { id } = request.params;
+
     const {
       name, email, phone, category_id,
     } = request.body;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid User Id' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
+    }
+
+    if (!name) {
+      return response.status(404).json({ error: 'name is required' });
+    }
+
     const contactExists = await ContactsRepository.findById(id);
 
     if (!contactExists) {
-      return response.status(404).json({ message: 'contact not found' });
+      return response.status(404).json({ error: 'contact not found' });
     }
-    if (!name) {
-      return response.status(404).json({ message: 'name is required' });
-    }
-    const contactEmailExists = await ContactsRepository.findByEmail(email);
 
-    if ((contactEmailExists && contactEmailExists.id !== id)) {
-      return response.status(404).json({ message: 'This e-mail address is already in use' });
+    if (email) {
+      const contactEmailExists = await ContactsRepository.findByEmail(email);
+
+      if ((contactEmailExists && contactEmailExists.id !== id)) {
+        return response.status(404).json({ error: 'This e-mail address is already in use' });
+      }
     }
     const contact = await ContactsRepository.update(id, {
-
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     response.status(200).json(contact);
@@ -67,6 +96,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid User Id' });
+    }
 
     await ContactsRepository.delete(id);
     response.sendStatus(204);
